requirements.txt
google-genai
edge-tts
genanki
requests
duckduckgo-search
phonemizer
python-dotenv
streamlit
watchdog
